requirements.txt
kivy>=2.3.0
pillow>=9.0.0
pyinstaller>=5.0.0
kivy_deps.sdl2
kivy_deps.glew 